Project notes
The impetus. This is what sparked this notebook.

A desire to work on small side projects with minimal overhead and an
easy way to take notes on what I was doing and working on.

No more getting bogged down by configuring a whole Repo and env so I can
write some code snippets/markdown notes.

Just a way to quickly and easily write about what I am doing.